Clean up blob and file
