+++ crates/file/src/blob.rs
@@ -1,46 +1,54 @@
-pub use crate::file::File;
-
 use wasm_bindgen::UnwrapThrowExt;
 
-pub trait Blob: raw::Blob {
+pub trait BlobLike {
     fn size(&self) -> usize {
-        self.raw().size() as usize
+        self.as_raw().size() as usize
     }
 
     fn mime_type(&self) -> Option<mime::Mime> {
-        Some(self.raw().type_().parse().ok()?)
+        Some(self.as_raw().type_().parse().ok()?)
     }
+
+    fn as_raw(&self) -> &web_sys::Blob;
 }
 
-pub struct DataBlob {
+pub struct Blob {
     inner: web_sys::Blob,
 }
 
-impl DataBlob {
-    pub fn new(content: &str) -> DataBlob {
+impl Blob {
+    pub fn new(content: &str) -> Blob {
         let parts = js_sys::Array::of1(&wasm_bindgen::JsValue::from_str(content));
         let inner = web_sys::Blob::new_with_str_sequence(&parts).unwrap_throw();
-        DataBlob { inner }
+        Blob { inner }
+    }
+
+    pub fn from_raw(inner: web_sys::Blob) -> Blob {
+        Blob { inner }
     }
 }
 
-impl Blob for DataBlob {}
-impl Blob for File {}
+impl BlobLike for Blob {
+    fn as_raw(&self) -> &web_sys::Blob {
+        &self.inner
+    }
+}
+impl BlobLike for File {
+    fn as_raw(&self) -> &web_sys::Blob {
+        &self.inner
+    }
+}
 
-mod raw {
-    pub trait Blob {
-        fn raw(&self) -> &web_sys::Blob;
+pub struct File {
+    pub(crate) inner: web_sys::File,
+}
+
+impl File {
+    pub fn from_raw(inner: web_sys::File) -> File {
+        File { inner }
     }
 
-    impl Blob for super::File {
-        fn raw(&self) -> &web_sys::Blob {
-            &self.inner
-        }
-    }
-
-    impl Blob for super::DataBlob {
-        fn raw(&self) -> &web_sys::Blob {
-            &self.inner
-        }
+    pub fn as_raw(&self) -> &web_sys::File {
+        &self.inner
     }
 }
