+++ crates/file/src/lib.rs
@@ -1,9 +1,7 @@
 mod blob;
-mod file;
 mod file_list;
 mod file_reader;
 
-pub use blob::{Blob, DataBlob};
-pub use file::File;
+pub use blob::{Blob, BlobLike, File};
 pub use file_list::FileList;
 pub use file_reader::FileReader;
